:lipstick: :recycle: Modularizado de componente de imagen con complejidades de graficado de funciones, y formateado en pares de imagen funcion en tarjetas
